+++ test/track-field-dependency-test.js
@@ -1,15 +1,15 @@
 import assert from 'assert';
 import Query from '../src/query';
 import typeBundle from '../fixtures/types'; // eslint-disable-line import/no-unresolved
-import {resetTracker, startTracking, pauseTracking, captureProfile} from '../src/track-type-dependency';
+import {resetProfiler, startProfiling, pauseProfiling, captureProfile} from '../src/profile-schema-usage';
 
 suite('track-field-dependency-test', () => {
   setup(() => {
-    resetTracker();
+    resetProfiler();
   });
 
   test('it reports the fields used in a query', () => {
-    startTracking();
+    startProfiling();
 
     // eslint-disable-next-line no-new
     new Query(typeBundle, (root) => {
@@ -66,8 +66,8 @@ suite('track-field-dependency-test', () => {
     });
   });
 
-  test('it pauses tracking when `pauseTracking` is called', () => {
-    startTracking();
+  test('it pauses tracking when `pauseProfiling` is called', () => {
+    startProfiling();
 
     // eslint-disable-next-line no-new
     new Query(typeBundle, (root) => {
@@ -76,7 +76,7 @@ suite('track-field-dependency-test', () => {
       });
     });
 
-    pauseTracking();
+    pauseProfiling();
 
     // eslint-disable-next-line no-new
     new Query(typeBundle, (root) => {
@@ -91,7 +91,7 @@ suite('track-field-dependency-test', () => {
       });
     });
 
-    startTracking();
+    startProfiling();
 
     // eslint-disable-next-line no-new
     new Query(typeBundle, (root) => {
@@ -117,9 +117,9 @@ suite('track-field-dependency-test', () => {
   });
 
   test('it stops tracking when `resetTypes` is called (returning the tracker to it\'s initial state.', () => {
-    startTracking();
+    startProfiling();
 
-    resetTracker();
+    resetProfiler();
 
     // eslint-disable-next-line no-new
     new Query(typeBundle, (root) => {
